e-center/requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
Pillow==11.0.0
django-filter==24.3
dj-database-url==2.2.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
whitenoise==6.8.2
coreapi==2.3.3
drf-spectacular==0.28.0
faker==24.4.0
django-extensions==3.2.3
google-generativeai==0.4.1

# Production dependencies (optional)
# redis==5.0.1
# celery==5.3.4
# sentry-sdk==1.38.0